# Use string type for current token and fingerprint