--- a/sqli_helpers.go
+++ b/sqli_helpers.go
@@ -111,8 +111,8 @@ func toUpperCmp(a, b string) bool {
 	return a == strings.ToUpper(b)
 }
 
-func isKeyword(key []byte) byte {
-	return searchKeyword(string(key), sqlKeywords)
+func isKeyword(key string) byte {
+	return searchKeyword(key, sqlKeywords)
 }
 
 func searchKeyword(key string, keywords map[string]byte) byte {