- `string` for current token avoids copying, it is always just a slice into the input
- Most logic use string operations so using a string for both reduces intermediate conversions to string

After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   32025	     35962 ns/op	   49568 B/op	     297 allocs/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	   12946	     93955 ns/op	   69889 B/op	    1718 allocs/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    6902	    165803 ns/op	  149562 B/op	    3689 allocs/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   29260	     40245 ns/op	   65665 B/op	     320 allocs/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	   12258	     98190 ns/op	   89521 B/op	    1841 allocs/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    6506	    179106 ns/op	  191738 B/op	    4136 allocs/op
PASS
```