@jcchavezs 2022-09-13
Merging this as:
1. changes look great
2. performance improved
3. Repo has good coverage level

If anything should change we can take care in a next PR.